+++ chats/apps/api/v1/rooms/serializers.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+from django.utils import timezone
 from rest_framework import serializers
 
 from chats.apps.accounts.models import User
@@ -104,7 +105,7 @@ class ListRoomSerializer(serializers.ModelSerializer):
     is_24h_valid = serializers.BooleanField(
         default=True, source="is_24h_valid_computed"
     )
-
+    config = serializers.JSONField(required=False, read_only=True)
     last_interaction = serializers.DateTimeField(read_only=True)
     can_edit_custom_fields = serializers.SerializerMethodField()
     is_active = serializers.BooleanField(default=True)
@@ -128,6 +129,7 @@ class ListRoomSerializer(serializers.ModelSerializer):
             "protocol",
             "service_chat",
             "is_active",
+            "config",
         ]
 
     def get_user(self, room: Room):
@@ -283,3 +285,41 @@ class RoomInfoSerializer(serializers.ModelSerializer):
             return first_user_message.created_on
 
         return None
+
+
+class RoomsReportFiltersSerializer(serializers.Serializer):
+    """
+    Filters for the rooms report.
+    """
+
+    created_on__gte = serializers.DateTimeField(required=True)
+    created_on__lte = serializers.DateTimeField(required=False)
+
+    def validate(self, attrs):
+        created_on__gte = attrs.get("created_on__gte")
+        created_on__lte = attrs.get("created_on__lte")
+
+        if created_on__gte and created_on__lte:
+            if created_on__gte > created_on__lte:
+                raise serializers.ValidationError(
+                    "created_on__gte must be before created_on__lte"
+                )
+
+        start = created_on__gte
+        end = created_on__lte or timezone.now()
+
+        period = (end - start).days
+
+        if period > 90:
+            raise serializers.ValidationError("Period must be less than 90 days")
+
+        return super().validate(attrs)
+
+
+class RoomsReportSerializer(serializers.Serializer):
+    """
+    Serializer for the rooms report.
+    """
+
+    recipient_email = serializers.EmailField(required=True)
+    filters = RoomsReportFiltersSerializer(required=True)