+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 3.18.11
+## Add
+  - new feature to see what messages are responded in chats.
+  - endpoint to generate rooms reports, generated asynchronously and sent via email
+
+# 3.18.10
+## Add
+  - billing notification when creating new rooms.
+  - new config field in room serializer
+
+# 3.18.9
+## Add
+  - Cache in ModuleHasPermission, to decrease the volume of calls to the database to check weather a internal user has the can_communicate_internally flag,
+  as this shouldn't change often.
+  - Increase the time for internal users (only) access token cache from 10 minutes to 6 hours.
+
 # 3.18.8
 ## Add
   - New field in room serializer, config.